requirements.txt
zhipuai>=1.0.7
matplotlib>=3.5.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
zai-sdk
python-dotenv>=0.19.0